fix watch no spec resources
