+++ pkg/utils/meta.go
@@ -1,7 +1,11 @@
 package utils
 
 import (
+	"context"
+
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/util/retry"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 // LabelsField presents a "f:labels" field subfield of metadataField.
@@ -62,4 +66,55 @@ func AddAnnotations(obj metav1.Object, annotations map[string]string) {
 	}
 
 	obj.SetAnnotations(mergedAnnotations)
+}
+
+// Remove is used to remove string from a string array
+func Remove(list []string, s string) []string {
+	result := []string{}
+	for _, v := range list {
+		if v != s {
+			result = append(result, v)
+		}
+	}
+	return result
+}
+
+// Contains is used to check whether a list contains string s
+func Contains(list []string, s string) bool {
+	for _, v := range list {
+		if v == s {
+			return true
+		}
+	}
+	return false
+}
+
+// GetAnnotation returns the annotation value for a given key, or an empty string if not set
+func GetAnnotation(annotations map[string]string, key string) string {
+	if annotations == nil {
+		return ""
+	}
+	return annotations[key]
+}
+
+func RemoveDuplicates(elements []string) []string {
+	// Use map to record duplicates as we find them.
+	encountered := map[string]struct{}{}
+	result := []string{}
+
+	for _, v := range elements {
+		if _, found := encountered[v]; found {
+			continue
+		}
+		encountered[v] = struct{}{}
+		result = append(result, v)
+	}
+	// Return the new slice.
+	return result
+}
+
+func UpdateObject(ctx context.Context, runtimeClient client.Client, obj client.Object) error {
+	return retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		return runtimeClient.Update(ctx, obj, &client.UpdateOptions{})
+	})
 }
